Boot celluloid in tests
For now this was required for me to get the tests running on the master
branch without Celluloid complaining that the thread pool had not been
started. I am not sure how this was working previously? Celluloid offers
some `autostart` or `test` requires, but neither of them seemed to work.